pyproject.toml
[tool.poetry]
name = "scheduler"
version = "0.1.0"
description = ""
authors = ["Alexander <sasha.14999933@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.38"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
jinja2 = "^3.1.5"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
